requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
statsmodels>=0.14
matplotlib>=3.7
seaborn>=0.12
nibabel>=5.0
tqdm>=4.65
python-docx>=0.8.11
